# pkg/models/debug_session.go
package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// DebugSession debug session
// swagger:model DebugSession
type DebugSession struct {

	// debug config Id
	// Required: true
	DebugConfigID *string `json:"debugConfigId"`

	// url
	URL string `json:"url,omitempty"`
}

// Validate validates this debug session
func (m *DebugSession) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateDebugConfigID(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *DebugSession) validateDebugConfigID(formats strfmt.Registry) error {

	if err := validate.Required("debugConfigId", "body", m.DebugConfigID); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *DebugSession) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *DebugSession) UnmarshalBinary(b []byte) error {
	var res DebugSession
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}